Add files via upload
The last one is a zip folder contains everything above. Still can't find a proper background for welcome page. No hyperlink linked right now. I upload one of my assignment so that you could use it as part of the register page if you would like. Also, I have no idea how to make the login page (to make it functional).  And I don't know how to make the comment system in Assignment 8, which we probably will use for that project.